Fix nyc compiler problems.
