+++ src/client/common/application/customEditorService.ts
@@ -4,22 +4,15 @@
 import { inject, injectable } from 'inversify';
 import * as vscode from 'vscode';
 
-import { ICommandManager, ICustomEditorService } from './types';
+import { ICommandManager, ICustomEditorService, WebviewCustomEditorProvider } from './types';
 
 @injectable()
 export class CustomEditorService implements ICustomEditorService {
     constructor(@inject(ICommandManager) private commandManager: ICommandManager) {}
 
-    public get supportsCustomEditors(): boolean {
-        try {
-            return vscode.window.registerWebviewCustomEditorProvider !== undefined;
-        } catch {
-            return false;
-        }
-    }
-
-    public registerWebviewCustomEditorProvider(viewType: string, provider: vscode.WebviewCustomEditorProvider, options?: vscode.WebviewPanelOptions): vscode.Disposable {
-        return vscode.window.registerWebviewCustomEditorProvider(viewType, provider, options);
+    public registerWebviewCustomEditorProvider(viewType: string, provider: WebviewCustomEditorProvider, options?: vscode.WebviewPanelOptions): vscode.Disposable {
+        // tslint:disable-next-line: no-any
+        return (vscode.window as any).registerWebviewCustomEditorProvider(viewType, provider, options);
     }
 
     public async openEditor(file: vscode.Uri): Promise<void> {
